+++ Models/Mods/GetModsByIdsListRequestBody.cs
@@ -4,6 +4,6 @@ namespace CurseForge.APIClient.Models.Mods
 {
     public class GetModsByIdsListRequestBody
     {
-        public List<long> ModIds { get; set; } = new List<long>();
+        public List<int> ModIds { get; set; } = new List<int>();
     }
 }
